requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tqdm>=4.62.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
jupyter>=1.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
albumentations>=1.3.0
tensorboard>=2.13.0
pillow>=9.0.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
memory-profiler>=0.60.0
optuna>=3.0.0
torch-tb-profiler>=0.4.0
wandb>=0.15.0
py-spy>=0.3.0
torch-pruning>=0.2.0
graphviz>=0.20.0
plotly>=5.13.0
psutil>=5.9.0 